% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlvm-package.R
\docType{package}
\name{bdlvm-package}
\alias{bdlvm}
\alias{bdlvm-package}
\title{bdlvm: Extend Other Libraries For Bayesian Latent Variable Modeling}
\description{
Wrappers for `brms`, `SBC`, and some extra convenience functions.
}
\keyword{internal}
