% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{bdlvm_parse}
\alias{bdlvm_parse}
\title{Parse a latent variable formula into \code{brms} format}
\usage{
bdlvm_parse(x, ...)
}
\arguments{
\item{x}{A formula created with \code{\link[=lv]{lv()}}}

\item{...}{Additional parameters to customize item naming and use of censoring/missingness. Currently supports \code{suffix1}, \code{suffix2}, \code{mi}, \code{cens}.}

\item{eval}{Default \code{TRUE}, otherwise returns the deparsed \code{bf()} calls}
}
\value{
\code{brms} formulas representing each latent variable and their corresponding measurement items
}
\description{
Constructs a set of \code{brms} formulas that encode latent variables using \code{mi()}
}
\examples{
library(bdlvm)
bdlvm_parse(lv(y ~ items(x, 3)))
}
