% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{lv}
\alias{lv}
\title{Set up a latent variable formula for \strong{brms}}
\usage{
lv(...)
}
\arguments{
\item{...}{Unnamed arguments must be formulas, see details

Any named argument is assumed to be a valid \code{\link[brms:brmsformula]{brms::bf()}} argument}
}
\value{
An S4 object of class \code{bdlvmFormula}, which supports concatenation via \code{+} with \code{brms} formulas.
}
\description{
Set up an object that can be parsed by \code{\link[=bdlvm]{bdlvm()}} into a \code{brms} formula that constructs a latent variable and the associated manifest variables through which it's measured.
}
\details{
The first RHS term must be of the form \code{items(prefix, num)}, where \code{num} is an integer and \code{prefix} is an unquoted name that will be used to build the measurement variables, following the pattern \code{prefixLVi1} up to \verb{prefixLVi\{num\}}. Additional terms on the RHS are interpreted as predictors on the latent variable's location parameter, as usual.

Subsequent formulas are interpreted as regressions on other parameters of the latent variable. See \code{brms} documentation for the list of supported families and their corresponding parameter names.
}
\examples{
# Single latent variable with three items (i.e. CFA)
lv(x ~ items(y, 3))

# Distributional regression, additional predictors and concatenation
# with brms formula. Note that latent variable 'x' does not require
# special notation to be placed in bf().
library(brms)
lv(x ~ items(y, 5) + z1,
   sigma ~ z1,
   family = "lognormal") +
   bf(z2 ~ x + z1)

}
