% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{parse_bdlvm}
\alias{parse_bdlvm}
\title{Parse a latent variable formula into \code{brms} format}
\usage{
parse_bdlvm(x, ...)
}
\arguments{
\item{x}{A formula created with \code{\link[=lv]{lv()}}}
}
\value{
A set of \code{brms} formulas corresponding with \code{mi()} terms added as needed.
}
\description{
Parse a latent variable formula into \code{brms} format
}
\examples{
library(bdlvm)
parse_bdlvm(lv(y ~ items(x, 3)))
}
