% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{parse_bdlvm}
\alias{parse_bdlvm}
\title{Parse a latent variable formula into \code{brms} format}
\usage{
parse_bdlvm(x, ...)
}
\arguments{
\item{x}{A formula created with \code{\link[=lv]{lv()}}}

\item{eval}{Default \code{TRUE}, otherwise returns the deparsed \code{bf()} calls}
}
\value{
\code{brms} formulas representing each latent variable and their corresponding measurement items
}
\description{
Constructs a set of \code{brms} formulas that encode latent variables using \code{mi()}
}
\examples{
library(bdlvm)
parse_bdlvm(lv(y ~ items(x, 3)))
}
